refactor: convert file from JSX to TSX for TypeScript support
